django-with-react-v2/requirements.txt
django~=3.1.1
djangorestframework~=3.11.2

